update path 9

+++ src/events/events.controller.ts
@@ -1,4 +1,4 @@
-import { Controller,Logger, Get, Post, Put, Delete, Patch, Param, NotFoundException,Body, HttpCode,ParseIntPipe,Res, ValidationPipe ,UsePipes,Query} from '@nestjs/common';
+import { Controller,Logger, Get, Post, Put, Delete, Patch, Param, NotFoundException,Body, HttpCode,ParseIntPipe,Res, ValidationPipe ,UsePipes,Query, UseGuards, ForbiddenException} from '@nestjs/common';
 import { CreateEventDto } from './create-event.dto';
 import { UpdateEventDto } from './update-event.dto';
 import { Event } from './event.entity';
@@ -7,16 +7,15 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Attendee } from './attendee.entity';
 import { EventsService } from './events.service';
 import { ListEvents } from './input/list.events';
+import { CurrentUser } from 'src/auth/current-user.decorator';
+import { User } from 'src/auth/user.entity';
+import { AuthGuardJwt } from 'src/auth/auth-guard.jwt';
 
 @Controller('/events')
 export class EventsController {
     private readonly logger = new Logger(EventsController.name)
 
     constructor(
-        @InjectRepository(Event)
-        private readonly repository: Repository<Event>,
-        @InjectRepository(Attendee)
-        private readonly attendeeRepository: Repository<Attendee>,
         private readonly eventsService:EventsService
     ) { }
 
@@ -27,45 +26,45 @@ export class EventsController {
         return events;
     }
 
-    @Get('/practice')
-    async practice() {
-        return await this.repository.find({
-            select:['id','when'],
-            where: [{ 
-                id: MoreThan(3) ,
-                when:MoreThan(new Date('2021-02-12T13:00:00'))
-            },{
-                description:Like('%meet%')
-            }],
-            take:2 ,//LIMIT
-            order:{ //Order By
-                id:'DESC'
-            }
-        });
-    }
+    // @Get('/practice')
+    // async practice() {
+    //     // return await this.repository.find({
+    //     //     select:['id','when'],
+    //     //     where: [{ 
+    //     //         id: MoreThan(3) ,
+    //     //         when:MoreThan(new Date('2021-02-12T13:00:00'))
+    //     //     },{
+    //     //         description:Like('%meet%')
+    //     //     }],
+    //     //     take:2 ,//LIMIT
+    //     //     order:{ //Order By
+    //     //         id:'DESC'
+    //     //     }
+    //     // });
+    // }
 
     
-    @Get('practice2/:id')
-    async  practice2(@Param('id',ParseIntPipe) id){
-        // return await this.repository.findOne({where:{id},relations:['attendees']})
-        const event = await this.repository.findOne({where:id,relations:['attendees']})
+    // @Get('practice2/:id')
+    // async  practice2(@Param('id',ParseIntPipe) id){
+    // //     // return await this.repository.findOne({where:{id},relations:['attendees']})
+    // //     const event = await this.repository.findOne({where:id,relations:['attendees']})
 
-        // const event = new Event()
-        // event.id=id
+    // //     // const event = new Event()
+    // //     // event.id=id
 
-        const attendee = new Attendee();
-        attendee.name = 'Using cascade'
-       // attendee.event = event
+    // //     const attendee = new Attendee();
+    // //     attendee.name = 'Using cascade'
+    // //    // attendee.event = event
 
-        event.attendees.push(attendee)
-        //event.attendees =[]
+    // //     event.attendees.push(attendee)
+    // //     //event.attendees =[]
 
-        // await this.attendeeRepository.save(attendee)
-        await this.repository.save(event)
+    // //     // await this.attendeeRepository.save(attendee)
+    // //     await this.repository.save(event)
 
-        return event
+    // //     return event
 
-    }
+    // }
 
 
     @Get(':id')
@@ -81,40 +80,47 @@ export class EventsController {
 
 
     @Post()
-    async create(@Body() input: CreateEventDto) {
-        await this.repository.save({
-            ...input,
-            when: new Date(input.when),
-        })
-
+    @UseGuards(AuthGuardJwt)
+    async create(@Body() input: CreateEventDto,
+    @CurrentUser() user:User
+    ) {
+       return await this.eventsService.createEvent(input,user)
 
     }
 
     @Patch(':id')
-    async update(@Param('id') id, @Body() input: UpdateEventDto) {
-        const event = await this.repository.findOne(id)
+    @UseGuards(AuthGuardJwt)
+    async update(@Param('id') id, @Body() input: UpdateEventDto,@CurrentUser() user:User) {
+        const event = await this.eventsService.getEvent(id)
 
         if(!event){
             throw new NotFoundException()
         }
 
-        return await this.repository.save({
-            ...event,
-            ...input,
-            when: input.when ? new Date(input.when) : event.when
-        })
+        if(event.organizerId !== user.id){
+            throw new ForbiddenException(null,`You are not authorized to change this event`)
+        }
+
+        return await this.eventsService.updateEvent(event,input)
 
 
     }
 
     @Delete(':id')
+    @UseGuards(AuthGuardJwt)
     @HttpCode(204)
-    async remove(@Param('id') id) {
-        const result = await this.eventsService.deleteEvent(id)
-        
-        if(result?.affected!==1){
+    async remove(@Param('id') id,@CurrentUser() user:User) {
+        const event = await this.eventsService.getEvent(id)
+
+        if(!event){
             throw new NotFoundException()
         }
-        
+
+        if(event.organizerId !== user.id){
+            throw new ForbiddenException(null,`You are not authorized to change this event`)
+        }
+
+        await this.eventsService.deleteEvent(id)
+  
     }
 }
